refactored dbdict.py to db_access.py
